Renaming files to better suit html5 structure
