+++ bigbluebutton-html5/imports/ui/components/actions-bar/component.jsx
@@ -4,7 +4,7 @@ import { styles } from './styles.scss';
 import EmojiSelect from './emoji-select/component';
 import ActionsDropdown from './actions-dropdown/component';
 import AudioControlsContainer from '../audio/audio-controls/container';
-import JoinVideoOptionsContainer from '../video-dock/video-menu/container';
+import JoinVideoOptionsContainer from '../video-provider/video-menu/container';
 
 class ActionsBar extends React.PureComponent {
   render() {
