+++ bigbluebutton-html5/imports/ui/components/video-provider/container.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { withTracker } from 'meteor/react-meteor-data';
 import VideoProvider from './component';
-import VideoService from '../video-dock/service';
+import VideoService from './service';
 
 const VideoProviderContainer = ({ children, ...props }) => <VideoProvider {...props}>{children}</VideoProvider>;
