+++ bigbluebutton-html5/imports/ui/components/video-provider/service.js
