+++ bigbluebutton-html5/imports/ui/components/video-provider/video-element/component.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { styles } from './styles';
+import { styles } from '../styles';
 
 class VideoElement extends Component {
   constructor(props) {
