+++ bigbluebutton-html5/imports/ui/components/video-provider/video-menu/service.js
